Total Rewrite of RhythmFinder Function
Total Rewrite of the function and HUGE simplification
Found a super helpful sequence on OEIS that related to the
task at hand.  After looking in to it found a formula for the
legnth of the pattern I was trying to search for in the number
line.  This allowed me to total ditch the pattern recognition
loops I had previously used in the code.  The Runtime is still
somewhere around N factorial. but now it shoudl be slightly
less than N! as opposed to the previous code, which was slightly
worse than N!.